Development upload #1
Changes:
- Working on Stream-based communications
- Working on vision system
- Working on electronics interfacing for Sbc
- Changed architecture of CommInterface - 
	- StreamCommInterface: abstract for stream-based communications
	- PortCommInterface: abstract for bus or serial communications
- Changed Pre-made actions to use less memory when running
- PidController can now work without WPILib
- Removed CommInfo as data class
- Added reflection and communications utilities to FlashUtil
- Added CommandParser
- Added CodeTester
- Added class instance tracker
- Added bandwidth tracker

+++ flashlibj/src/edu/flash3388/flashlib/robot/actions/VisionAction.java
@@ -4,9 +4,9 @@ import edu.flash3388.flashlib.robot.systems.ModableMotor;
 import edu.flash3388.flashlib.vision.Vision;
 
 public interface VisionAction {
-	public static final long LOSS_TIMEOUT = 200;
-	public static final int ACCURACY_MARGIN = 10;
-	public static final long ACTION_VALIDATION_TIMEOUT = 50;
+	public static final int LOSS_TIMEOUT = 200;
+	public static final byte ACCURACY_MARGIN = 10;
+	public static final int ACTION_VALIDATION_TIMEOUT = 50;
 	public static final double DEFAULT_MIN_SPEED = 0.15;
 	public static final double DEFAULT_MAX_SPEED = 0.8;
 	
@@ -24,6 +24,6 @@ public interface VisionAction {
 	double getMinSpeed();
 	void setMaxSpeed(double max);
 	double getMaxSpeed();
-	long getLossTimeout();
-	void setLossTimeout(long millis);
+	int getLossTimeout();
+	void setLossTimeout(int millis);
 }
